% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable.R
\name{crosstable}
\alias{crosstable}
\title{Generate a table for cross tabulations}
\usage{
crosstable(data, x, y, x_names = FALSE, y_names = FALSE)
}
\arguments{
\item{data}{Object holding the dataframe}

\item{x}{String of the first variable name}

\item{y}{String of the second variable name}

\item{x_names}{Vector of response options. Default value is FALSE.}

\item{y_names}{Vector of response options. Default value is FALSE.}
}
\value{
A flextable object
}
\description{
This function generates a flextable for a cross tabulation
between two categorical (nominal or ordinal), numeric variables within a dataframe.
There is an option to make the column names of the table reflect response options
instead of the default, where the column names are the actual numeric values.
}
\examples{
mat <- as.data.frame(matrix(1:20, 5, 4, dimnames = list(NULL, LETTERS[1:4])))
crosstable(mat, "A", "B")
crosstable(mat, "A", "B", x_names = c("one", "two", "three", "four", "five"), y_names = c("six", "seven", "eight", "nine", "ten"))
}
