% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable_ext.R
\name{crosstable_ext}
\alias{crosstable_ext}
\title{Generates a table for cross tabulations with percentages}
\usage{
crosstable_ext(
  data,
  x,
  y,
  x_names = FALSE,
  y_names = FALSE,
  percents = "TRC",
  row_t = TRUE,
  col_t = TRUE
)
}
\arguments{
\item{data}{Object holding the dataframe}

\item{x}{String of the first variable name}

\item{y}{String of the second variable name}

\item{x_names}{Vector of response options. Default value is FALSE.}

\item{y_names}{Vector of response options. Default value is FALSE.}

\item{percents}{String representing what percentages to calculate. Accepts any of c("TRC", "RC", "TC", "TR", "C", "R", "T", "N"). Default is "TRC".}

\item{row_t}{Boolean determining if row totals should be included. Default value is TRUE.}

\item{col_t}{Boolean determining if column totals should be included. Default value is TRUE.}
}
\value{
A flextable object
}
\description{
This function generates a flextable for a cross tabulation
between two categorical (nominal or ordinal), numeric variables within a dataframe.
The cell values contains the total percentage, column percentage, row percentage, and count.
There is an option to make the column names of the table reflect response options
instead of the default, where the column names are the actual numeric values.
}
\examples{
mat <- as.data.frame(matrix(1:20, 5, 4, dimnames = list(NULL, LETTERS[1:4])))
crosstable_ext(mat, "A", "B")
crosstable_ext(mat, "A", "B", x_names = c("one", "two", "three", "four", "five"), y_names = c("six", "seven", "eight", "nine", "ten"))
}
