% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex2df.R
\name{flex2df}
\alias{flex2df}
\title{Convert a flextable object to a dataframe}
\usage{
flex2df(flex, tbl_type)
}
\arguments{
\item{flex}{A flextable object}

\item{tbl_type}{String referring to type of table. Accepts "unitable" or "crosstable".}
}
\value{
A data.frame object
}
\description{
This function takes a flextable object as input and returns a data.frame object
containing the content of the flextable cells.
}
