% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanBoxplot.R
\name{meanBoxplot}
\alias{meanBoxplot}
\title{Generate a ggplot2 box and whisker plot using mean and standard deviations}
\usage{
meanBoxplot(data, x_var, y_var)
}
\arguments{
\item{data}{Variable containing the data.frame}

\item{x_var}{A categorical variable for the x axis}

\item{y_var}{A continuous or semi-continuous variable for the y axis}
}
\value{
A ggplot2 graph object
}
\description{
This function takes two columns from a data.frame object and generates a ggplot2 object. The
ggplot2 object is a box and whisker plot where the median line is replaced with the mean, the
first and third quartile box is replaced with the first standard deviation above and below the mean,
and the whiskers are replaced with the minimum and maximum values of the data. Variables containing
the output of this function can be treated as the root for further ggplot2 additions (ex. labs,
geom_jitter).
}
