% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitable.R
\name{unitable}
\alias{unitable}
\title{Generate a table for a single variable}
\usage{
unitable(
  data,
  colval,
  minval = NA,
  maxval = NA,
  namecol = NA,
  text = " ",
  title = NA
)
}
\arguments{
\item{data}{Object holding the dataframe}

\item{colval}{Call to the variable in the dataframe formatted as df$colname}

\item{minval}{Optional. Integer representing the smallest possible numeric value in the variable}

\item{maxval}{Optional. Integer representing the largest possible numeric value in the variable}

\item{namecol}{Optional. Vector containing string of response options.}

\item{text}{Optional. String of the question wording in the survey}

\item{title}{Optional. String of the table title, usually the variable name in the dataframe}
}
\value{
A flextable object
}
\description{
This function generates a flextable for the distribution of responses to a
single categorical (nominal or ordinal) numeric variable of a dataframe.
}
\examples{
mat <- as.data.frame(matrix(1:20, 5, 4, dimnames = list(NULL, LETTERS[1:4])))
unitable(mat, mat$A, 1, 5, c("A", "one", "two", "three", "four", "five"), "What is the question?", 1, "Example table")
}
